Added a C++ Plugin to the project but Unity does not yet interface with it correctly
Created an Android Studio project with "Include C++ Support" ticked,
built the project, and copied the “.so” files that were generated into
“Assets/Plugins/Android/libs/“.

In AWSS3Client I’m then attempting to import a simple function I wrote
that just returns an int, however the “DllImport” directive is failing
with the following output:  “Unable to lookup library path for
'native-lib', native render plugin support disabled.”

The idea is to attempt to use a C++ Plugin to run the Texture2D
processing on a seperate thread. But first we need the plugin to simply
work!