C++ Plugin is now being correctly picked up
The problem was that the default ".so" files that are generated from
the Android Studio 2.2 project are called "libnative-lib.so".

For whatever reason, Unity's toolset doesn’t recognise them as a plugin
due to their hyphen, once I removed the hyphen and called it
"libandroidcppnative.so” then Unity recognised it.

Then with the new name it got added correctly to the apk when I built
it!

Now we need to write the actual C++ code that will run on a background
thread and generate the Texture2D without blocking.