Part-way through work on C++ Plugin
With the plugin connecting up from C++ to Unity I started working on
getting the plugin to actually perform the Image processing that it
needs to do.

First of all I managed to import the stb_image.h library correctly and
include it without a hitch into my C++ code.

With the library at hand, I managed to correctly send the S3 image
stream buffer into the plugin and calculate the dimensions of the
image, which in turn I used to create a new Texture2D.

I then managed to run a new thread in C# calling my C++ code
successfully. So far so good!

However, when it came to having this seperate thread run a function
that would internally call “stbi_load_from_memory”, my luck ran out.

For whatever reason running this function, and setting the result back
into the Texture2D I created, does not seem to work, and putting
printf’s into the C++ code is not showing up in adb logcat making it
hard to debug the problem.