Running the Image decoding on a background thread!
We’re now very close to fixing the problem! We’ve managed to fix the
following issues:

(1) The Plugin now succeeds in downloading the image (although for some
reason the image is upside down).
(2) The work done in the plugin runs on a background thread using the
new class ThreadJob, which the AWSS3Client utilises.

In order to fix the problem entirely we’ll still need to:

(1) Get the image loading the correct way up.
(2) Remove small stutters that still occur (which I believe are caused
by the memory being read from the ResponseStream and SetPixels() being
called).