Tidied up AWSS3Client code to a certain extent
I moved the converting from Stream to byte[] onto a background thread,
and diagnosed that unfortunately the stutter that I’m still getting is
when the Texture2D is created. I’m not quite sure how (if at all) I’ll
be able to get around this.

When trying to get the image facing the correct way up I attempted to
simply invert the byte[], however that crashed the program when trying
to Load it.

Therefore, I still need to figure out how to flip the image the correct
way up, and if its possible to move the Texture2D creation to another
time in the running of the program…