Fixed the Upside Down Image issue!
Cleon fixed this issue by removing the memcpy in the C++ plugin code,
and replacing it with a for-loop using pointer arithmetic to juggle the
memory around.

I had been going about it through the C# side, naively attempting to
reverse the byte[], but this solution was the more correct, and better
way to approach it.

Now we just need to remove the stutter on creation of the Texture2D,
which Cleon also has an idea for solving!