Added an FPS display + Moved plugin code into a wrapper class called CppPlugin
The FPS counter is a simple script I grabbed online, and just helps
make it really obvious when we’ve framed out.

I’ve created a CppPlugin class which wraps the plugin code so that
there’s a centralised place from which to access and run the plugin
related functions!