Removed the y-axis flip when Loading images
In my TransferPixelsFromSrcToDest() function I split the whole process
of scanning through the pixels into width-long scan lines, then went
through each scan line in reverse order.

With regards to reading memory this means that we now read from the
very bottom upwards (as we were doing before), but then we minus the
width of a scan line and read forwards towards the end of the scan line!