Attempting to remove the Texture2D Construction and its stall on the main thread - not yet succeeded
I’ve written a new LoadImageFromPath(), which is essentially the same
as the previous one but I’m now calling
Texture2D.CreateExternalTexture() and handing it a pointer to some
Marshalled memory (which won’t be GC’d).

This isn’t working though and currently everything I try simply returns
a black image. It feels like this method is close to the correct way to
solve the problem, but there’s something key that I’m missing somewhere
along the way. =/