Managed to get the C++ Plugin to Create a new correctly sized texture every time!
Its now the case that the Texture is not being pre-fabricated in the C# side and passed to C++ for filling in, but instead the entire Texture creation and pixel manipulation is happening on the C++ Plugin side of things without interference from C# at all. The only thing that C# is doing is (1) setting the ImagePath, (2) giving the order for the C++ plugin to do its thing, (3) calling CreateExternalTexture(), then (4) setting it on the Sphere.

This is huge progress! The main next thing I need to do is have this code actually run without stalling the Main Thread!