Its DONE! Textures are loading in through the C++ Plugin Asynchronously and not causing any noticeable performance dips!
I’ve now actually got a small set of “kNumTextures” textures under the hood that I’m cycling over and refreshing with the contents of the new image before Creating a new Texture2D in C#!

The plugin can currently only handle the loading of one texture at a time, and what it does is it utilises a set of state variables (m_currTextureIndex, m_currImageWidth, m_currImageHeight) that are valid for the Texture currently being loaded/worked on, such that the C# code can then query it and be certain that it knows what’s happening to the texture in C++ land.

I know need to integrate this code into my actual main project!