Tidied up Texture2DTest as a whole!
Removed the AWS S3 code, because it muddied the actual point of the Test repo, which is to solve non-blocking loading of Texture2Ds.

Similarly removed images being on spheres, because that also gets in the way of the repo’s main purpose.

Its now the case that the loading happens on a flat plane, and I don’t hardcode the path so it can find any image on the users device.

I’ve also removed any mention of the JavaPlugin because it wasn’t used in the end.